+++ astropy/coordinates/builtin_frames/baseradec.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-from ... import units as u
 from .. import representation as r
 from ..baseframe import BaseCoordinateFrame, RepresentationMapping
 
@@ -58,29 +57,8 @@ class BaseRADecFrame(BaseCoordinateFrame):
         r.SphericalRepresentation: [
             RepresentationMapping('lon', 'ra'),
             RepresentationMapping('lat', 'dec')
-        ],
-        r.SphericalCosLatDifferential: [
-            RepresentationMapping('d_lon_coslat', 'pm_ra_cosdec', u.mas/u.yr),
-            RepresentationMapping('d_lat', 'pm_dec', u.mas/u.yr),
-            RepresentationMapping('d_distance', 'radial_velocity', u.km/u.s)
-        ],
-        r.SphericalDifferential: [
-            RepresentationMapping('d_lon', 'pm_ra', u.mas/u.yr),
-            RepresentationMapping('d_lat', 'pm_dec', u.mas/u.yr),
-            RepresentationMapping('d_distance', 'radial_velocity', u.km/u.s)
-        ],
-        r.CartesianDifferential: [
-            RepresentationMapping('d_x', 'v_x', u.km/u.s),
-            RepresentationMapping('d_y', 'v_y', u.km/u.s),
-            RepresentationMapping('d_z', 'v_z', u.km/u.s)
-        ],
+        ]
     }
-    frame_specific_representation_info[r.UnitSphericalRepresentation] = \
-        frame_specific_representation_info[r.SphericalRepresentation]
-    frame_specific_representation_info[r.UnitSphericalCosLatDifferential] = \
-        frame_specific_representation_info[r.SphericalCosLatDifferential]
-    frame_specific_representation_info[r.UnitSphericalDifferential] = \
-        frame_specific_representation_info[r.SphericalDifferential]
 
     default_representation = r.SphericalRepresentation
     default_differential = r.SphericalCosLatDifferential